% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLPSscore.R
\name{getLPSscore}
\alias{getLPSscore}
\title{LPS score calculation for a given sample}
\usage{
getLPSscore(vdat, coefs)
}
\arguments{
\item{vdat}{a numeric vector for all selected features for a given sample}

\item{coefs}{a numeric vector of weights for all selected features, 
which should be the same order as in "vdat"}
}
\value{
A LPS score
}
\description{
This is an internal function to calculate LPS (Linear Prediction Score),
 which is called by other functions but it can also be called directly.
}
\details{
LPS calculation is based on Wright 2003, the formula is staightforward:
 \eqn{LPS(X) = \sum a_j X_j}
If NAs are not imputed, they are ignored for LPS calculation.
}
\references{
Wright G, Tan B, Rosenwald A, Hurt EH, Wiestner A, Staudt LM. A gene expression-based method
to diagnose clinically distinct subgroups of diffuse large B cell lymphoma. Proc Natl Acad Sci U S
A. 2003 Aug 19;100(17):9991-6.
}
\author{
Aixiang Jiang
}
\keyword{LPS}
