% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanOfGroupMeans.R
\name{getMeanOfGroupMeans}
\alias{getMeanOfGroupMeans}
\title{Mean of two groups' means calculation}
\usage{
getMeanOfGroupMeans(traitdat, groupInfo, refGroup = 0)
}
\arguments{
\item{traitdat}{a vector of data for a single feature}

\item{groupInfo}{a known group classification, which sample order should be the same as for traitdat}

\item{refGroup}{the code for reference group, default is 0, but it can be a string or other number, 
which will be changed to 0 within the function}
}
\value{
A numerical vector of mean of two group means, and the two group means
}
\description{
This is an internal function called by PStraining to calculate mean of two groups' means for a feature,
 which is needed for PS appraoch. However, it can be called directly if you do not have PStraining object but want to calcualate
 PS scores, in this case, however, please make sure that you have get subset of selected features data only as traitdat, and
 provide groupInfo and refGroup, then, the output of this function can be used to get PS scores via getClassScores function.
}
\details{
When mean of each group feature data is calculated, the NAs are removed.
}
\author{
Aixiang Jiang
}
\keyword{mean}
\keyword{sd}
