% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainingWeights.R
\name{getTrainingWeights}
\alias{getTrainingWeights}
\title{Feature selection and their weight calculation for PS, LPS and PRPS approaches}
\usage{
getTrainingWeights(trainDat, selectedTraits = NULL, groupInfo,
  refGroup = 0, topN = NULL, FDRcut = 0.1,
  weightMethod = c("ttest", "limma", "PearsonR", "SpearmanR",
  "MannWhitneyU"))
}
\arguments{
\item{trainDat}{training data set, a data matrix or a data frame, samples are in columns, 
and features/traits are in rows}

\item{selectedTraits}{a selected trait list if available}

\item{groupInfo}{a known group classification, which order should be the same as in colnames of trainDat}

\item{refGroup}{the code for reference group, default is 0, but it can be a string or other number, 
which will be changed to 0 within the function}

\item{topN}{an integer to indicate how many top features to be selected}

\item{FDRcut}{FDR cutoff for feature selection}

\item{weightMethod}{a string to indicate weight calculation method, there are five choices: 
"limma" for for limma linear model based t value,"ttest" for t test based t value, 
"MannWhitneyU" for Mann Whitney U based rank-biserial,"PearsonR" for Pearson correlation coefficient,
"SpearmanR" for Spearman correlation coefficient, and the defualt value is "ttest"}
}
\value{
A data frame of selected features' parameters such as weights and FDR
}
\description{
this is a wrap up function to select traits, and get their weights for each selected feature
}
\details{
Within the five weight calculation methods, "limma","ttest", "PearsonR","SpearmanR" are often used and nothing 
is changed for them from usual setting in this function. We describe more detail about "MannWhitneyU" here. 
Mann-Whitney U test is a well known rank based two group comparison test method, but its statistics value U is not easy 
to use for weight. U can be calculated in two different ways depending on which group we stand for. 
Assuming that there are two groups with sample sizes n1 and n2, and rank sum for the two groups are R1 and R2, 
then we can have the following two U values:
\eqn{ U_1 = R_1 - n_1(n_1 + 1)/2}
\eqn{ U_2 = R_2 - n_2(n_2 + 1)/2}
In this function, we use the test group (non reference group) to calculate U value, and then calculate 
its rank-biserial correlation with the following formula:
\eqn{ r = 1 - 2U/(n_1 * n_2)}
P value for "MannWhitneyU", however, is still based on standard Mann - Whitney U test.
NA is ignored before testing.
}
\references{
Smyth, G. K. (2004). Linear models and empirical Bayes methods for assessing differential expression in microarray experiments. 
Statistical Applications in traittics and Molecular Biology, Volume 3, Article 3. http://www.statsci.org/smyth/pubs/ebayes.pdf
Wendt, H.W. (1972). "Dealing with a common problem in social science: A simplified rank-biserial coefficient of correlation 
based on the U statistic". European Journal of Social Psychology. 2 (4): 463–465. doi:10.1002/ejsp.2420020412.
}
\author{
Aixiang Jiang
}
\keyword{feature}
\keyword{selection}
\keyword{weight}
