% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraitParsWithEM.R
\name{getTraitParsWithEM}
\alias{getTraitParsWithEM}
\title{Calculate two group means, sds, and mean of group means for each of features, which are used in PRPS and PS}
\usage{
getTraitParsWithEM(datin, weights)
}
\arguments{
\item{datin}{a new data matrix or data frame, columns for samples and rows for features}

\item{weights}{a numeric vector with selected features (as names of the vector) and their weights}
}
\value{
A data frame with group means, sds and mean of group means
}
\description{
This function is to calculate feature level group mean, sd and mean of means, without knowing sample group label.
}
\details{
In order to get feature level group means, sd sand mean of means, without knowing sample group label, we use EM algorithm to 
decompose a vector of values into two normal distributions, and obtain parameters (mean and sd) for each of them, 
and then get mean of group means. This is done for each given feature. In order to determine the direction for each 
feature, we do need weight for each selected feature.
}
\references{
Ultsch, A., Thrun, M.C., Hansen-Goos, O., Loetsch, J.: Identification of Molecular Fingerprints
in Human Heat Pain Thresholds by Use of an Interactive Mixture Model R Toolbox(AdaptGauss),
International Journal of Molecular Sciences, doi:10.3390/ijms161025897, 2015.
}
\author{
Aixiang Jiang
}
\keyword{EM}
\keyword{mean}
\keyword{sd}
