% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNAs.R
\name{imputeNAs}
\alias{imputeNAs}
\title{NA value imputation}
\usage{
imputeNAs(dataIn, byrow = TRUE, imputeValue = c("median", "mean"))
}
\arguments{
\item{dataIn}{a data matrix or a data frame, samples are in columns, and features/genes are in rows}

\item{byrow}{although we recommend to impute NAs along with features, but this function can be used for either direction,
the default value is byrow = TRUE}

\item{imputeValue}{for classification, the most non-informatic value is median/mean in a given row/column, 
so there are two choices for the imputed values: median or mean when NAs are excluded, the default value is "median"}
}
\value{
Same data matrix or data frame as input except that the NA values are imputed.
}
\description{
This function is to impute NA values if there are any NA in the data. 
Without this step, NAs will cause a lot of troubles in classification score calculation, and therefore, we have to ignore NAs.
}
\details{
During the classification score calculation process, NA offen causes troubles especially when matrix multiplication is involved. 
To avoid this problem, we suggest to impute NA values before any other functions are called if you have NA in your data. 
For classification purpose, the median value is believed to be most noninformative value, so we use median to replace NA. Further more, 
we prefer to do feature-wise median replacement although this function can be applied in both feature wise and sample wise replacements.
}
\author{
Aixiang Jiang
}
\keyword{NA}
\keyword{impute}
