% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPSSLextension.R
\name{LPSSLextension}
\alias{LPSSLextension}
\title{LPS score calculation and binary classification for a testing new data set based on a self learning object}
\usage{
LPSSLextension(LPSSLObj, newdat, standardization = FALSE,
  classProbCut = 0.8, imputeNA = FALSE, byrow = TRUE,
  imputeValue = c("median", "mean"))
}
\arguments{
\item{LPSSLObj}{a LPS self learning object, which is the output from either LPSSLwithWeightsEM or LPSSLwithWeightsPrior}

\item{newdat}{a new data matrix or data frame, which is comparable to self learning data set, 
with columns for samples and rows for features}

\item{standardization}{a logic variable to indicate if standardization is needed before classification 
score calculation}

\item{classProbCut}{a numeric variable within (0,1), which is a cutoff of Empirical Bayesian probability, 
often used values are 0.8 and 0.9, default value is 0.8. Only one value is used for both groups, 
the samples that are not included in either group will be assigned as UNCLASS}

\item{imputeNA}{a logic variable to indicate if NA imputation is needed, if it is TRUE, NA imputation is 
processed before any other steps, the default is FALSE}

\item{byrow}{a logic variable to indicate direction for imputation, default is TRUE, 
which will use the row data for imputation}

\item{imputeValue}{a character variable to indicate which value to be used to replace NA, default is "median", 
the median value of the chose direction with "byrow" data to be used}
}
\value{
A data frame with LPS score, Empirical Bayesian probabilites for two groups and classification
}
\description{
This is the function to calculate LPS (Linear Prediction Score) scores for a testing data set 
with LPS self learning object. The selected feature list, these features' parameters are from the given LPS self learning object
that is an output from either LPSSLwithWeightsEM or LPSSLwithWeightsPrior.
}
\details{
This is the function to calculate LPS scores and make classification and 
Empirical Bayesian probabilities for a testing new data set. However, this new data set should be 
comparable to the self learning data set as much as possible. Within LPS self learning and within this current 
functions, standardization step is included as an option to minimize the difference 
between self learning and testing data sets, but this step is only done to make distributions of each selected
features comparable. Be aware that this feature-wise standardization cannot make the sample-wise distributions
comparable. For example, the self learning data set must have two classification groups, however, 
the proportion of one group sample might be much less than the other group in the testing data set 
compared to the self learning data set, or even worse, the testing data set might only contain one classification 
group only. This is the common problem for classification and feature-wise standardization cannot solve the problem. 
In order to solve the problem, we should make data comparable as much as possbile before classification step. 
For example, use the same pre-processing settings and make suitable batch effect correction. 

LPS calculation is based on Wright 2003. The fomula is straightforward:
  \eqn{LPS(X) = \sum a_j x_ij}
Here a_j represents the jth selected feature weights, and x_ij is the corresponding feature value for the ith sample.
LPS mean and sd for two groups are needed for Empirical Bayes' probability calculation, which are from a LPS 
self learning object that is an output from either LPSSLwithWeightsEM or LPSSLwithWeightsPrior.

When calculate a Empirical Bayes' probability, the 1st group in the input mean and sd vectors is treated as the 
test group. When calculate the probabilities, we first calcualte probability that a sample belongs to either group,
and then use the following formula to get Empirical Bayes' probability:
\eqn{prob(x) = p_test(x)/(p_test(x) + p_ref(x))}
Here prob(x) is the Empirical Bayes' probability of a given sample, p_test(x) is the probability that a given sample
belongs to the test group, p_ref(x) is the probability that a given sample belongs to the reference group.
Notice that the test and reference group is just the relative grouping, in fact, for this step, we often need
 to calculate Empirical Bayes' probabilities for a given sample from two different standing points.
}
\references{
Wright G, Tan B, Rosenwald A, Hurt EH, Wiestner A, Staudt LM. A gene expression-based method
to diagnose clinically distinct subgroups of diffuse large B cell lymphoma. Proc Natl Acad Sci U S
A. 2003 Aug 19;100(17):9991-6.
}
\author{
Aixiang Jiang
}
\keyword{LPS}
