% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{ROC curve with AUC, Youden best and natural cutoff of 0}
\usage{
plotROC(contdat, contname, catdat, catname, xshift = -0.05,
  yshift = 0.02)
}
\arguments{
\item{contdat}{a vector of a continuous variable}

\item{contname}{a string to show name of a continuous vector}

\item{catdat}{a vector of a dichotomous variable}

\item{catname}{a string to show name of a  dichotomous variable}

\item{xshift}{a numeric variable to indicate how much value shift along x-axis to move the label position for the natural 0 cutoff point,
default value is}

\item{yshift}{a numeric variable to indicate how much value shift along y-axis to move the label position for the natural 0 cutoff point,
default value is}
}
\description{
This is an internal function to plot ROC (Receiver operating characteristic) curve showing AUC, Youden best and natural cutoff of 0 points.
}
\author{
Aixiang Jiang
}
\keyword{ROC}
