% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraining.R
\name{plotTraining}
\alias{plotTraining}
\title{Plot function for training objects}
\usage{
plotTraining(trainObj, plotNanme, xshift, yshift)
}
\arguments{
\item{trainObj}{a training object from LPStrain, or PRPStrain, or PStrain}

\item{xshift}{a numeric variable to indicate how much value shift along x-axis to move the label position for the natural 0 cutoff point,
default value is}

\item{yshift}{a numeric variable to indicate how much value shift along y-axis to move the label position for the natural 0 cutoff point,
default value is}
}
\description{
This is to plot ROC, hist, and scatter plot for a given training object
}
\author{
Aixiang Jiang
}
\keyword{ROC,}
\keyword{hist,}
\keyword{plot}
\keyword{scatter}
