% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRPSClassWithWeights.R
\name{PRPSClassWithWeight}
\alias{PRPSClassWithWeight}
\title{PRPS score calculation and binary classification for a testing data set without PRPStraining output object, but with selected feature weights}
\usage{
PRPSClassWithWeight(newdat, weights, standardization = FALSE,
  classProbCut = 0.8, PRPShighGroup = "PRPShigh",
  PRPSlowGroup = "PRPSlow", breaks = 50, imputeNA = FALSE,
  byrow = TRUE, imputeValue = c("median", "mean"))
}
\arguments{
\item{newdat}{a new data matrix or data frame, columns for samples and rows for features}

\item{weights}{a numeric vector with selected features (as names of the vector) and their weights}

\item{standardization}{a logic variable to indicate if standardization is needed before classification 
score calculation}

\item{classProbCut}{a numeric variable within (0,1), which is a cutoff of Empirical Bayesian probability, 
often used values are 0.8 and 0.9, default value is 0.8. Only one value is used for both groups, 
the samples that are not included in either group will be assigned as UNCLASS}

\item{PRPShighGroup}{a string to indicate group name with high PRPS score}

\item{PRPSlowGroup}{a string to indicate group name with low PRPS score}

\item{breaks}{a integer to indicate number of bins in histogram, default is 50}

\item{imputeNA}{a logic variable to indicate if NA imputation is needed, if it is TRUE, NA imputation is 
processed before any other steps, the default is FALSE}

\item{byrow}{a logic variable to indicate direction for imputation, default is TRUE, 
which will use the row data for imputation}

\item{imputeValue}{a character variable to indicate which value to be used to replace NA, default is "median", 
the median value of the chose direction with "byrow" data to be used}
}
\value{
A data frame with PRPS score, Empirical Bayesian probabilites for two groups and classification
}
\description{
This is the function to calculate PRPS (Probability ratio based classification predication score) scores for a testing data set 
without PRPS training object. However, we do need selected feature list with their weights and mean and sd for two groups for each
selected feature, which we need to apply EM (Expectation-Maximization) to achieve. Once we have PRPS scores, we also need 
to apply EM to calcualate mean and sd for the two groups assuming that PRPS score is a mixture of two normal distributions, 
followed by Empirical Bayes' probability calculation and final binary classification calls.
}
\details{
This is the function to calculate PRPS scores and make classification based on
Empirical Bayesian probabilities for a testing new data set. 
PRPS calculation is based on Ennishi 2018, its formula is:
\eqn{LPS(X_i) = \sum (|a_j| log(P1(x_ij)/P0(x_ij)))}
Here, a_j represents the jth selected feature weights, and x_ij is the corresponding feature value
 for the ith sample, 
P1 and P0 are the probabilities that the ith sample belongs to two different group.
However, in order to calculate P1 and P0, we need to have two group mean and sd for each selected feature. Although there are multiple way
to obtain these values, in this function, we design to use EM algorithm to achieve group mean and sd assuming that each selected feature
is a mixture of two normal distributions. 

After we have PRPS scores, we need to calculate a Empirical Bayes' probability to make classification calls. To do that, 
we also need to apply EM to get PRPS score mean and sd for two groups. 
After that, we can calcualte probability that a sample belongs to either group,
and then use the following formula to get Empirical Bayes' probability:
\eqn{prob(x) = p_test(x)/(p_test(x) + p_ref(x))}
Here prob(x) is the Empirical Bayes' probability of a given sample, p_test(x) is the probability that a given sample
belongs to the test group, p_ref(x) is the probability that a given sample belongs to the reference group.
Notice that the test and reference group is just the relative grouping, in fact, for this step, we often need
 to calculate Empirical Bayes' probabilities for a given sample from two different standing points.
}
\references{
Wright G, Tan B, Rosenwald A, Hurt EH, Wiestner A, Staudt LM. A gene expression-based method
to diagnose clinically distinct subgroups of diffuse large B cell lymphoma. Proc Natl Acad Sci U S
A. 2003 Aug 19;100(17):9991-6.

Ultsch, A., Thrun, M.C., Hansen-Goos, O., Loetsch, J.: Identification of Molecular Fingerprints
in Human Heat Pain Thresholds by Use of an Interactive Mixture Model R Toolbox(AdaptGauss),
International Journal of Molecular Sciences, doi:10.3390/ijms161025897, 2015.
}
\author{
Aixiang Jiang
}
\keyword{EM}
\keyword{PRPS}
