% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{A standardization Function}
\usage{
standardize(dataIn, byrow = TRUE)
}
\arguments{
\item{dataIn}{a data matrix or a data frame, samples are in columns, and features/genes are in rows}

\item{byrow}{although we recommend to standardize data along with features, 
but this function can be used for either direction, the default value is byrow = TRUE}
}
\value{
Standardized data matrix or data frame
}
\description{
This function is to standardize data in order to make features or samples be comparable. 
You can to standardize your data either by row or by column.
}
\details{
Standardization is the standard way to make data comparable. For this package, 
this should be done right after imputeNAs if necessary. In general, 
we do not recommend this step before classification, however, if the training and testing data sets
are obviously not comparable, we suggest to standardize training and testing data separately for each gene. 
It should be aware that this step might inflate sample differences for the low expression genes.
Standardization is done as:
 \eqn{ X = (X-mean(X))/sd(X) }
In this step, NA are removed from mean and sd calculation, which means that this step is safe even imputeNAs is not done before this step when they are NAs in the data.
}
\author{
Aixiang Jiang
}
\keyword{standardization}
