% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraining.R
\name{plotTraining}
\alias{plotTraining}
\title{Plot function for training objects}
\usage{
plotTraining(trainObj, plotName, breaks = 30)
}
\arguments{
\item{trainObj}{a training object from LPStraining, or PRPStraining, or PStraining}

\item{breaks}{a integer to indicate how many cells in the histogram}

\item{plotNanme}{a string variable to indicate the file name to save, it should includes path and plot file name (without .pdf part)}
}
\description{
This is to plot hist for a given training object, and scatter plot if the training object us from LPStraining or PRPStraining
}
\author{
Aixiang Jiang
}
\keyword{hist,}
\keyword{plot}
\keyword{scatter}
