% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRPSstableSLwithWeights.R
\name{PRPSstableSLwithWeights}
\alias{PRPSstableSLwithWeights}
\title{Self learning binary classification with selected features and their weights}
\usage{
PRPSstableSLwithWeights(newdat, weights, standardization = FALSE,
  classProbCut = 0.9, PRPShighGroup = "PRPShigh",
  PRPSlowGroup = "PRPSlow", breaks = 50, EMmaxRuns = 50,
  imputeNA = FALSE, byrow = TRUE, imputeValue = c("median", "mean"))
}
\arguments{
\item{newdat}{a input data matrix or data frame, columns for samples and rows for features}

\item{weights}{a numeric vector with selected features (as names of the vector) and their weights}

\item{standardization}{a logic variable to indicate if standardization is needed before classification 
score calculation}

\item{classProbCut}{a numeric variable within (0,1), which is a cutoff of Empirical Bayesian probability, 
often used values are 0.8 and 0.9, default value is 0.9. Only one value is used for both groups, 
the samples that are not included in either group will be assigned as UNCLASS}

\item{PRPShighGroup}{a string to indicate group name with high PRPS score}

\item{PRPSlowGroup}{a string to indicate group name with low PRPS score}

\item{breaks}{a integer to indicate number of bins in histogram, default is 50}

\item{EMmaxRuns}{number of Iterations for EM searching; default=50}

\item{imputeNA}{a logic variable to indicate if NA imputation is needed, if it is TRUE, NA imputation is 
processed before any other steps, the default is FALSE}

\item{byrow}{a logic variable to indicate direction for imputation, default is TRUE, 
which will use the row data for imputation}

\item{imputeValue}{a character variable to indicate which value to be used to replace NA, default is "median", 
the median value of the chose direction with "byrow" data to be used}
}
\value{
A list with two items is returned: PRPS parameters for selected features, PRPS scores and classifications for the given samples.
\item{PRPS_pars}{a list of 3 items, the 1st item is a data frame with weights of each selected features for PRPS
 calculation, the 2nd item is a numeric vector containing PRPS mean and sd for two groups，the 3rd item is a data frame contains mean and sd
  for each group and for each selected feature based on stable classes}
\item{PRPS_test}{a data frame of PRPS score, classification and two groups' Empirical Bayesian probabilites based on stable classes, 
and classification with natural 0 cutoff}
}
\description{
This function is to calculate PRPS (Probability ratio based classification predication score) scores and 
make binary classificationcalls for a testing data set without PRPS training object. This function involves a self learning 
process with weights + priors + EM + Bayes but no need to input a group ratio prior. However, we do need selected feature list
with the feature weights, and we use self learning method to estimate mean and sd for two groups for each selected feature 
in order to calculate PRPS scores. Our original idea in JCO 2018 does requrire a group ratio prior to do so,
which is written as PRPSSLwithWeightsPrior function in this package. We also have another function: PRPSSLwithWeightsEM, 
which uses EM to estimate group mean and sd for each selected feature. In this fucnction, however, we combine the ideas in both
of the above two functions to achieve our new version of self learning algorithm.
}
\details{
This function is trying to get reasonable PRPS based classification without training data set, but with 
selected features and their weights. It combines ideas from PRPSSLwithWeightsPrior and
PRPSSLwithWeightsEM, which we call it as self learning algorithm, the actual steps are as following:
1) assume that we have a pool for group ratio priors such as: seq(0.05, 0.95, by = 0.05), this will give us 19 ratio priors
2) for each prior in 1), call PRPSSLwithWeightsPrior to achieve PRPS scores
3) apply EM on PRPS scores from 2) with Mclust, which includes 2 group classification
4) use the samples that are always in the same groups to get group means and sds for each feature and for each group
5) calculate PRPS scores
6) Once we have PRPS scores, we could use the theoretic natual cutoff 0 to make classification calls, which may or may not appropriate. 
Alternatively, we can also apply EM to calcualate mean and sd for the
two groups based on stable classes assuming that PRPS score is a mixture of two normal distributions, followed by Empirical Bayes' probability.
}
\references{
Ennishi D, Jiang A, Boyle M, Collinge B, Grande BM, Ben-Neriah S, Rushton C, Tang J, Thomas N, Slack GW, Farinha P, 
 Takata K, Miyata-Takata T, Craig J, Mottok A, Meissner B, Saberi S, Bashashati A, Villa D, Savage KJ, Sehn LH, Kridel R, 
 Mungall AJ, Marra MA, Shah SP, Steidl C, Connors JM, Gascoyne RD, Morin RD, Scott DW. Double-Hit Gene Expression Signature Defines
 a Distinct Subgroup of Germinal Center B-Cell-Like Diffuse Large B-Cell Lymphoma. J Clin Oncol. 
 2018 Dec 3:JCO1801583. doi: 10.1200/JCO.18.01583.

Ultsch, A., Thrun, M.C., Hansen-Goos, O., Loetsch, J.: Identification of Molecular Fingerprints
in Human Heat Pain Thresholds by Use of an Interactive Mixture Model R Toolbox(AdaptGauss),
International Journal of Molecular Sciences, doi:10.3390/ijms161025897, 2015.

Scrucca L., Fop M., Murphy T. B. and Raftery A. E. (2016) mclust 5: clustering, classification and 
density estimation using Gaussian finite mixture models, The R Journal, 8/1, pp. 205-233.
}
\author{
Aixiang Jiang
}
\keyword{EM}
\keyword{PRPS}
\keyword{self-learning}
